chore!(autommands): moved to lua/user/autommands.lua
