+++ lua/options/init.lua
@@ -69,64 +69,3 @@ local buffer_options = {
 for key, value in pairs(buffer_options) do
   vim.bo[key] = value
 end
-
---
--- Autocommands
---
---
--- To reset window options that had been set by other autocmd. This is needed because win options
--- are set for the whole window, not only buffers, thus they carry over to other buffers. I'd be
--- nice to have these options available for buffers.
-local function reset_win_options()
-  vim.api.nvim_win_set_option(0, "colorcolumn", "100")
-  vim.api.nvim_win_set_option(0, "spell", false)
-  vim.api.nvim_win_set_option(0, "linebreak", false)
-end
-
-vim.api.nvim_create_autocmd("BufLeave", {
-  group = vim.api.nvim_create_augroup("reset win options", { clear = true }),
-  pattern = { "*" },
-  callback = reset_win_options,
-})
-
-vim.api.nvim_create_autocmd("FileType", {
-  group = vim.api.nvim_create_augroup("bufcheck", { clear = true }),
-  pattern = { "gitcommit", "gitrebase" },
-  command = "startinsert | 1",
-})
-
--- Python filetypes
-local function filetypes_python() vim.api.nvim_win_set_option(0, "colorcolumn", "80") end
-
-vim.api.nvim_create_autocmd({ "BufEnter" }, {
-  group = vim.api.nvim_create_augroup("filetype python", { clear = true }),
-  pattern = { "*.py" },
-  callback = filetypes_python,
-})
-
-vim.api.nvim_create_autocmd("TextYankPost", {
-  group = vim.api.nvim_create_augroup("highlight_on_yank", { clear = true }),
-  command = "lua vim.highlight.on_yank { higroup='IncSearch', timeout=750}",
-})
-
-vim.api.nvim_create_autocmd("InsertLeave", {
-  group = vim.api.nvim_create_augroup("leave_insert_set_nopaste", { clear = true }),
-  command = "silent! set nopaste",
-})
-
--- `documentation` filetypes
-local documentation_file_options = function()
-  vim.api.nvim_win_set_option(0, "conceallevel", 2)
-  vim.api.nvim_win_set_option(0, "spell", true)
-  vim.bo.fo = "awtq"
-  vim.wo.linebreak = true
-end
-
-vim.api.nvim_create_autocmd("BufEnter", {
-  group = vim.api.nvim_create_augroup("documentation_file_options", { clear = true }),
-  pattern = { "*.md", "*.tex" },
-  callback = documentation_file_options,
-})
-
-vim.cmd([[
-]])
